pyproject.toml
[tool.poetry]
name = "backend-python-template"
version = "1.0.0"
description = "Python FastAPI with PostgreSQL Database"
authors = ["Anuj Raval"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.25"
requests = "^2.31.0"
alembic = "^1.13.1"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
typing = "3.7.4.3"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
